Enemy and Enemy Type files merged together
