+++ Assets/Scenes/MainScene.unity
@@ -1589,11 +1589,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 835538455}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c811cae7c0ad9b3428469b44193e14fe, type: 3}
+  m_Script: {fileID: 11500000, guid: a477a54fff7a24c0c9a77c816c3d9d49, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   name: Test Enemy
-  health: 34
+  health: 30
+  spell: 0
+  isUnlocked: 1
 --- !u!1 &839270415
 GameObject:
   m_ObjectHideFlags: 0
