+++ Assets/Scripts/Enemy.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d4d5aaa65149b644d8a91312dbc0dd2e
+guid: 67fa28c90c37c4b56b051e55b5af7b2e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
