+++ Assets/Scripts/Enemy/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c811cae7c0ad9b3428469b44193e14fe
+guid: a477a54fff7a24c0c9a77c816c3d9d49
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
