+++ Assets/Scripts/Enemy/Sunflower.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 // Notice this class is a child of Enemy
 
-public class AppleTree : Enemy
+public class Sunflower : Enemy
 {
-    public AppleTree()
+    public Sunflower()
     {
-        name = "Apple Tree";
-        spell = 15;
-        costToKill = 15;
+        name = "Sunflower";
+        spell = 18;
+        health = 18;
         isUnlocked = true;
     }
 }
