+++ Assets/Scripts/Enemy/Watermelon.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 // Notice this class is a child of Enemy
 
-public class Corn : Enemy
+public class Watermelon : Enemy
 {
-    public Corn()
+    public Watermelon()
     {
-        name = "Corn";
-        spell = 5;
-        costToKill = 5;
+        name = "Watermelon";
+        spell = 20;
+        health = 20;
         isUnlocked = true;
     }
 }
