--- Assets/Scripts/EnemyTypes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 67fa28c90c37c4b56b051e55b5af7b2e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
