--- Assets/Scripts/EnemyTypes/Enemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a477a54fff7a24c0c9a77c816c3d9d49
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
